cnn-classifier/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==1.12.0
torchvision==0.13.0
numpy==1.24.4
Pillow==10.1.0
python-multipart==0.0.6
jinja2==3.1.2